+++ examples/resources/morpheus_max_hosts_policy/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_max_hosts_policy.tf_example_max_hosts_policy 1