+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * **New Data Source:** `morpheus_workflow`
 * **New Resource:** `morpheus_max_cores_policy`
+* **New Resource:** `morpheus_max_hosts_policy`
 * **New Resource:** `morpheus_max_vms_policy`
 * **New Resource:** `morpheus_workflow_policy`